+++ EngineSIU/EngineSIU/Engine/Source/Runtime/CoreUObject/UObject/Class.h
